fix(diffusion): restore Qwen QKV CI invariants
